Clean up a bit the rendering module, to prepare transition with FlatTree.
